feat. 배너 컴포넌트 생성
